それぞれの[id].tsxの中のcategoriesを作る処理をまとめた
